cloud_functions/map_github_taxonomy/requirements.txt
google-cloud-storage==2.17.0
